Tweak headers and generate cpucuda kernel launch

+++ clang/lib/Headers/cpucuda_detail/types.hpp
@@ -34,13 +34,13 @@ namespace cpucuda {
 
 	struct dim3
 	{
-		dim3(size_t d0 = 1, size_t d1 = 1, size_t d2 = 1)
+		dim3(unsigned d0 = 1, unsigned d1 = 1, unsigned d2 = 1)
 			: x{d0}, y{d1}, z{d2}
 		{}
 
-		size_t x;
-		size_t y;
-		size_t z;
+		unsigned x;
+		unsigned y;
+		unsigned z;
 	};
 
 }
